goal CRUD is set with user authentication :development

+++ backend/controller/goalController.js
@@ -1,11 +1,11 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
-const mn = require('mongoose')
+const User =  require('../models/userModel')
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async(req,res) =>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -20,6 +20,7 @@ const setGoals = asyncHandler(async(req,res) =>{
     }
    const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id
    })
    res.status(200).json(goal)
 })
@@ -33,11 +34,27 @@ const updateGoals = asyncHandler(async(req,res) =>{
         res.status(400)
         throw new Error('Goal not Found');
     }
-    await Goal.findByIdAndUpdate(req.params.id,req.body,{
+
+    const user = await User.findById(req.user.id)
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id.toString()){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{
         new: true
     })
-    const updatedgoal = await Goal.findById(req.params.id)
-    res.status(200).json(goal);  
+    if (!updatedGoal) {
+        res.status(400);
+        throw new Error('Failed to update goal');
+    }
+    res.status(200).json(updatedGoal);  
 })
 
 
@@ -45,11 +62,27 @@ const updateGoals = asyncHandler(async(req,res) =>{
 //@route DELETE /api/goals
 //@access Private
 const deleteGoals = asyncHandler(async(req,res) =>{
-    const goal = Goal.findById(req.params.id)
+    const goal = await Goal.findById(req.params.id)
+    console.log(goal);
     if(!goal){
         res.status(400)
         throw new Error('Goal not Found')
     }
+    
+    const user = await User.findById(req.user.id)
+    console.log(user);
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id.toString()){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     await Goal.findByIdAndDelete(req.params.id)
     res.status(200).json({id :req.params.id});
 })
